requirements.txt
fastapi==0.116.1
passlib==1.7.4
pydantic==2.11.7
pymongo==3.12.0
python-dotenv==1.1.1
python_jose==3.5.0
uvicorn[standard]==0.32.1
